requirements.txt
pandas
scikit-learn
matplotlib
seaborn
pyyaml
openpyxl
joblib
